+++ data/Pandora.Definitions.ResourceManager/LoadTestService/Terraform/LoadTest-Resource.cs
@@ -11,13 +11,18 @@ public class LoadTestResource : TerraformResourceDefinition
     public string ResourceLabel => "load_test";
     public string ResourceCategory => "Load Test";
     public string ResourceDescription => @"Manages a Load Test Service";
-    public string ResourceExampleUsage => @"provider 'azurerm' {
-  features {}
+    public string ResourceExampleUsage => @"resource 'azurerm_resource_group' 'example' {
+  name     = 'example-resources'
+  location = 'West Europe'
 }
-
+resource 'azurerm_user_assigned_identity' 'example' {
+  name                = 'example'
+  resource_group_name = azurerm_resource_group.example.name
+  location            = azurerm_resource_group.example.location
+}
 resource 'azurerm_load_test' 'example' {
   location            = azurerm_resource_group.example.location
-  name                = 'examplelt'
+  name                = 'example'
   resource_group_name = azurerm_resource_group.example.name
 }".AsTerraformTestConfig();
     public Type? SchemaModel => typeof(LoadTestResourceSchema);
