+++ data/Pandora.Definitions.ResourceManager/Sql/v2021_02_01_preview/ManagedDatabaseSensitivityLabels/Constant-SensitivityLabelSource.cs
@@ -1,7 +1,7 @@
 using Pandora.Definitions.Attributes;
 using System.ComponentModel;
 
-namespace Pandora.Definitions.ResourceManager.Sql.v2022_11_01_preview.SensitivityLabels;
+namespace Pandora.Definitions.ResourceManager.Sql.v2021_02_01_preview.ManagedDatabaseSensitivityLabels;
 
 [ConstantType(ConstantTypeAttribute.ConstantType.String)]
 internal enum SensitivityLabelSourceConstant
