+++ data/Pandora.Definitions.ResourceManager/Sql/v2021_02_01_preview/ManagedDatabaseSensitivityLabels/Definition.cs
@@ -27,6 +27,7 @@ internal class Definition : ResourceDefinition
     {
         typeof(RecommendedSensitivityLabelUpdateKindConstant),
         typeof(SensitivityLabelRankConstant),
+        typeof(SensitivityLabelSourceConstant),
         typeof(SensitivityLabelUpdateKindConstant),
     };
     public IEnumerable<System.Type> Models => new List<System.Type>
