+++ data/Pandora.Definitions.ResourceManager/Sql/v2021_02_01_preview/ManagedDatabaseSensitivityLabels/Operation-CreateOrUpdate.cs
@@ -16,11 +16,9 @@ internal class CreateOrUpdateOperation : Pandora.Definitions.Operations.PutOpera
 {
     public override Type? RequestObject() => typeof(SensitivityLabelModel);
 
-    public override ResourceID? ResourceId() => new TableColumnId();
+    public override ResourceID? ResourceId() => new SensitivityLabelSourceId();
 
     public override Type? ResponseObject() => typeof(SensitivityLabelModel);
-
-    public override string? UriSuffix() => "/sensitivityLabels/current";
 
 
 }
