+++ data/Pandora.Definitions.ResourceManager/Sql/v2021_02_01_preview/ManagedDatabaseSensitivityLabels/Operation-Get.cs
@@ -14,11 +14,9 @@ namespace Pandora.Definitions.ResourceManager.Sql.v2021_02_01_preview.ManagedDat
 
 internal class GetOperation : Pandora.Definitions.Operations.GetOperation
 {
-    public override ResourceID? ResourceId() => new TableColumnId();
+    public override ResourceID? ResourceId() => new SensitivityLabelSourceId();
 
     public override Type? ResponseObject() => typeof(SensitivityLabelModel);
-
-    public override string? UriSuffix() => "/sensitivityLabels/current";
 
 
 }
