--- data/Pandora.Definitions.ResourceManager/Sql/v2022_11_01_preview/ManagedDatabaseSensitivityLabels/Constant-SensitivityLabelSource.cs
@@ -1,14 +1,0 @@
-using Pandora.Definitions.Attributes;
-using System.ComponentModel;
-
-namespace Pandora.Definitions.ResourceManager.Sql.v2022_11_01_preview.ManagedDatabaseSensitivityLabels;
-
-[ConstantType(ConstantTypeAttribute.ConstantType.String)]
-internal enum SensitivityLabelSourceConstant
-{
-    [Description("current")]
-    Current,
-
-    [Description("recommended")]
-    Recommended,
-}
