+++ data/Pandora.Definitions.ResourceManager/Sql/v2022_11_01_preview/SensitivityLabels/Operation-CreateOrUpdate.cs
@@ -16,9 +16,11 @@ internal class CreateOrUpdateOperation : Pandora.Definitions.Operations.PutOpera
 {
     public override Type? RequestObject() => typeof(SensitivityLabelModel);
 
-    public override ResourceID? ResourceId() => new SensitivityLabelSourceId();
+    public override ResourceID? ResourceId() => new ColumnId();
 
     public override Type? ResponseObject() => typeof(SensitivityLabelModel);
+
+    public override string? UriSuffix() => "/sensitivityLabels/current";
 
 
 }
