--- data/Pandora.Definitions.ResourceManager/Sql/v2022_11_01_preview/SensitivityLabels/ResourceId-SensitivityLabelSourceId.cs
@@ -1,38 +1,0 @@
-using System.Collections.Generic;
-using Pandora.Definitions.Interfaces;
-
-
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
-
-
-namespace Pandora.Definitions.ResourceManager.Sql.v2022_11_01_preview.SensitivityLabels;
-
-internal class SensitivityLabelSourceId : ResourceID
-{
-    public string? CommonAlias => null;
-
-    public string ID => "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Sql/servers/{serverName}/databases/{databaseName}/schemas/{schemaName}/tables/{tableName}/columns/{columnName}/sensitivityLabels/{sensitivityLabelSource}";
-
-    public List<ResourceIDSegment> Segments => new List<ResourceIDSegment>
-    {
-        ResourceIDSegment.Static("staticSubscriptions", "subscriptions"),
-        ResourceIDSegment.SubscriptionId("subscriptionId"),
-        ResourceIDSegment.Static("staticResourceGroups", "resourceGroups"),
-        ResourceIDSegment.ResourceGroup("resourceGroupName"),
-        ResourceIDSegment.Static("staticProviders", "providers"),
-        ResourceIDSegment.ResourceProvider("staticMicrosoftSql", "Microsoft.Sql"),
-        ResourceIDSegment.Static("staticServers", "servers"),
-        ResourceIDSegment.UserSpecified("serverName"),
-        ResourceIDSegment.Static("staticDatabases", "databases"),
-        ResourceIDSegment.UserSpecified("databaseName"),
-        ResourceIDSegment.Static("staticSchemas", "schemas"),
-        ResourceIDSegment.UserSpecified("schemaName"),
-        ResourceIDSegment.Static("staticTables", "tables"),
-        ResourceIDSegment.UserSpecified("tableName"),
-        ResourceIDSegment.Static("staticColumns", "columns"),
-        ResourceIDSegment.UserSpecified("columnName"),
-        ResourceIDSegment.Static("staticSensitivityLabels", "sensitivityLabels"),
-        ResourceIDSegment.Constant("sensitivityLabelSource", typeof(SensitivityLabelSourceConstant)),
-    };
-}
